pyproject.toml
[project]
name = "record-website"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ffmpeg>=1.4",
    "ffmpeg-python>=0.2.0",
    "google-api-python-client>=2.175.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "gspread>=6.2.1",
    "playwright>=1.53.0",
]
